[Blob DB] Move BlobFile definition to a separate file
